https://www.shopusa.com/in/
SHIPPING COURIER FROM USA TO INDIA
Worldwide the obsessive attraction for products from USA retailers has shown a widespread increase and the same cannot be denied by anyone in India as well. The reasons for the attraction can be because of the fact that most of them are not available in India if at all they are they are too expensive and thirdly even if they are available the choices are very limited. Also if you buy them from online shopping services many times they do not have the options of shipping from USA to India. If at all they manage to, their charges, taxes, customs duties and the processes for customs clearance are so lengthy and tiring that it would mean buying them in triple the actual cost. The retailers themselves also cannot do the shipping to India because of the absence of export information, the heavy cost involved in courier from USA to India and then marketing them here. Also, the market size of USA is huge enough for them to elevate the business.